+++ multiqc/modules/bbmap/bbmap.py
@@ -2,7 +2,7 @@
 from __future__ import print_function
 import logging
 from collections import OrderedDict
-
+from multiqc.utils import config
 from multiqc.plots import table
 from multiqc.modules.base_module import BaseMultiqcModule
 
@@ -72,6 +72,29 @@ class MultiqcModule(BaseMultiqcModule):
                     helptext=file_types[file_type]["help_text"],
                     plot=self.make_basic_table(file_type),
                 )
+
+        # Special case - qchist metric in General Stats
+        if "qchist" in self.mod_data:
+            data = {}
+            fraction_gt_q30 = []
+            for s_name in self.mod_data["qchist"]:
+                for qual, d in self.mod_data["qchist"][s_name]["data"].items():
+                    if int(qual) >= 30:
+                        fraction_gt_q30.append(d[1])
+                data[s_name] = {"pct_q30": sum(fraction_gt_q30) * 100.0}
+
+            headers = {
+                "pct_q30": {
+                    "title": "% Q30 bases",
+                    "description": "BBMap qchist - Percentage of bases with phred quality score >= 30",
+                    "suffix": " %",
+                    "scale": "RdYlGn",
+                    "format": "{:,.2f}",
+                    "min": 0,
+                    "max": 100,
+                }
+            }
+            self.general_stats_addcols(data, headers)
 
     def parse_logs(self, file_type, root, s_name, fn, f, **kw):